+++ turbopack/crates/turbopack-tests/tests/execution/turbopack/scope-hoisting/new/input/auth2.js
@@ -1,4 +1,4 @@
-import { freeze } from './src/util/object-utils.js'
+import { freeze } from './src/utils.js'
 console.log(freeze)
-import { parseExpression } from './src/parser/expression-parser.js'
+import { parseExpression } from './src/parser.js'
 console.log(parseExpression)
