+++ turbopack/crates/turbopack-tests/tests/execution/turbopack/scope-hoisting/new/input/src/builder.js
@@ -1,0 +1,7 @@
+import { parseSelectAll } from './select.js'
+console.log(parseSelectAll)
+
+import { freeze } from './utils.js'
+console.log(freeze)
+
+export function expressionBuilder() {}
