--- turbopack/crates/turbopack-tests/tests/execution/turbopack/scope-hoisting/new/input/src/expression/expression-builder.js
@@ -1,7 +1,0 @@
-import { parseSelectAll } from '../parser/select-parser.js'
-console.log(parseSelectAll)
-
-import { freeze } from '../util/object-utils.js'
-console.log(freeze)
-
-export function expressionBuilder() {}
