+++ turbopack/crates/turbopack-tests/tests/execution/turbopack/scope-hoisting/new/input/src/parser.js
@@ -1,0 +1,4 @@
+import { expressionBuilder } from './builder.js'
+console.log(expressionBuilder)
+
+export function parseExpression() {}
