--- turbopack/crates/turbopack-tests/tests/execution/turbopack/scope-hoisting/new/input/src/parser/select-parser.js
@@ -1,4 +1,0 @@
-import { expressionBuilder } from '../expression/expression-builder.js'
-console.log(expressionBuilder)
-
-export function parseSelectAll() {}
